+++ frontend/index.tsx
@@ -2,7 +2,7 @@
 import { StrictMode } from 'react'
 import ReactDOM from 'react-dom/client'
 import './index.css'
-import App from './app'
+import App from './app-a'
 import TimeAgo from 'javascript-time-ago'
 import en from 'javascript-time-ago/locale/en'
